chore(shade):organized modules and imports
