+++ packages/shades/src/services/index.ts
@@ -1,0 +1,1 @@
+export * from './dom-observer-service'
